Allow a static build on Windows/MSVC using MSYS2 as build
It is not currently possible to build shared MSVC libraries
using libtool (either the one bundled with libffi 2.4.1 or
the latest from git).

The reason for this is because libtool aliases the name of
import libraries and the name of static libraries for that
toolchain (in this case, both are libffi.lib).

Instead, we need to patch libtool to use defacto standard
on Windows/MSVC of prefixing the static libraries with an
extra 'lib'. This is what Microsoft and also Boost does.

I will work on that when I can. notes.md is a shell script
that regenerates the patch to perform the libtool update
and should be re-run once the libtool work is finished.

For now, we pass --enable-static --disable-shared because
some libraries are better than none, right?